// src/app/shared/types/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.12 (cd3cf9e)"
  }
  public: {
    Tables: {
// ...
      gpu_benchmarks: {
        Row: {
          avg_pwr_draw: number | null
          avg_temp: number | null
          cuda_cores: number | null
          cuda_version: string | null
          d2d_mem_bandwidth: number | null
          d2h_mem_bandwidth: number | null
          driver_version: string | null
          fp16_flops: number | null
          fp32_flops: number | null
          fp64_flops: number | null
          gpu_aib_partner: string | null
          gpu_arch: string | null
          gpu_model: string | null
          h2d_mem_bandwidth: number | null
          id: number
          inserted_at: string
          max_pwr_draw: number | null
          max_temp: number | null
          num_sm: number | null
          rt_cores: number | null
          tensor_cores: number | null
          tensor_flops_bf16: number | null
          tensor_flops_fp16: number | null
          tensor_flops_fp8: number | null
          tensor_flops_int8: number | null
          tensor_flops_tf32: number | null
          updated_at: string
          user_id: string | null
          uuid: string
          vram_gb: number | null
        }
        Insert: {
          avg_pwr_draw?: number | null
          avg_temp?: number | null
          cuda_cores?: number | null
          cuda_version?: string | null
          d2d_mem_bandwidth?: number | null
          d2h_mem_bandwidth?: number | null
          driver_version?: string | null
          fp16_flops?: number | null
          fp32_flops?: number | null
          fp64_flops?: number | null
          gpu_aib_partner?: string | null
          gpu_arch?: string | null
          gpu_model?: string | null
          h2d_mem_bandwidth?: number | null
          id?: number
          inserted_at?: string
          max_pwr_draw?: number | null
          max_temp?: number | null
          num_sm?: number | null
          rt_cores?: number | null
          tensor_cores?: number | null
          tensor_flops_bf16?: number | null
          tensor_flops_fp16?: number | null
          tensor_flops_fp8?: number | null
          tensor_flops_int8?: number | null
          tensor_flops_tf32?: number | null
          updated_at?: string
          user_id?: string | null
          uuid: string
          vram_gb?: number | null
        }
        Update: {
          avg_pwr_draw?: number | null
          avg_temp?: number | null
          cuda_cores?: number | null
          cuda_version?: string | null
          d2d_mem_bandwidth?: number | null
          d2h_mem_bandwidth?: number | null
          driver_version?: string | null
          fp16_flops?: number | null
          fp32_flops?: number | null
          fp64_flops?: number | null
          gpu_aib_partner?: string | null
          gpu_arch?: string | null
          gpu_model?: string | null
          h2d_mem_bandwidth?: number | null
          id?: number
          inserted_at?: string
          max_pwr_draw?: number | null
          max_temp?: number | null
          num_sm?: number | null
          rt_cores?: number | null
          tensor_cores?: number | null
          tensor_flops_bf16?: number | null
          tensor_flops_fp16?: number | null
          tensor_flops_fp8?: number | null
          tensor_flops_int8?: number | null
          tensor_flops_tf32?: number | null
          updated_at?: string
          user_id?: string | null
          uuid?: string
          vram_gb?: number | null
        }
        Relationships: []
      }
      gpu_listings: {
        Row: {
          benchmark_id: number
          condition: string
          created_at: string | null
          description: string | null
          id: number
          image_urls: string[] | null
          location: string | null
          price_amount: number
          price_currency: string
          status: string
          title: string
          updated_at: string | null
          user_id: string | null
        }
        Insert: {
          benchmark_id: number
          condition: string
          created_at?: string | null
          description?: string | null
          id?: never
          image_urls?: string[] | null
          location?: string | null
          price_amount: number
          price_currency?: string
          status?: string
          title: string
          updated_at?: string | null
          user_id?: string | null
        }
        Update: {
          benchmark_id?: number
          condition?: string
          created_at?: string | null
          description?: string | null
          id?: never
          image_urls?: string[] | null
          location?: string | null
          price_amount?: number
          price_currency?: string
          status?: string
          title?: string
          updated_at?: string | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "gpu_listings_benchmark_id_fkey"
            columns: ["benchmark_id"]
            isOneToOne: false
            referencedRelation: "gpu_benchmarks"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "gpu_listings_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
      users: {
        Row: {
          email: string
          id: string
          phone_number: number
          username: string | null
        }
        Insert: {
          email: string
          id: string
          phone_number: number
          username?: string | null
        }
        Update: {
          email?: string
          id?: string
          phone_number?: number
          username?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      generate_gpu_benchmark_avgs: {
        Args: Record<PropertyKey, never>
        Returns: undefined
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never